Always take up either all of the height or all of the width
With a fixed Ratio you can run into a situation where we shrink the height of a row to account for wider elements. This change makes it that if you do that then other rows grow proportionally to take up the extra space. So you should always either take up the full height or the full width. There shouldn't be a situation where there is whitespace all the way around.