fixes #36 - Adding an alignItems option
Adding alignItems, bigAlignItems and smallAlignItems. The values can be 'start', 'center' or 'end' and this determines how to arrange items in a row that is not full. Either at the start, the center or the end. The small setting is just for when there is a big item and determines how to arrange the small row or column.